feat:自定义数据库镜像实现
1.基于mysql基镜像编写dockerfile，实现自己的mysql镜像（启动时自动创建blog数据库与相关table）
2.修改users table id数据类型，解决于file table文件关联类型不兼容冲突

+++ blog-backend/migrations/20241011_create_users_table.up.sql
@@ -17,7 +17,7 @@
 -- ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
 
 CREATE TABLE IF NOT EXISTS `users` (
-    `id` int NOT NULL AUTO_INCREMENT,
+    `id` BIGINT(20) NOT NULL AUTO_INCREMENT,
     `username` varchar(128) NOT NULL comment '用户登录名',
     `password` varchar(128) NOT NULL comment '用户密码',
     `role` tinyint DEFAULT '1' comment '角色',
